04_Professionnal__(Days82_to_Day100)/Day96__ReadMe_Analyzer/requirements.txt
flask-bootstrap5==0.1.dev1
requests==2.32.3
